+++ src/main/scala/org/fs/checker/cache/CachedDetails.scala
@@ -2,7 +2,11 @@ package org.fs.checker.cache
 
 import com.github.nscala_time.time.Imports._
 
-case class CachedDetails(lastCheckMsOption: Option[Long], lastUpdateMsOption: Option[Long]) {
+case class CachedDetails(
+  lastCheckMsOption:  Option[Long],
+  lastUpdateMsOption: Option[Long],
+  isUnavailable:      Boolean
+) {
   def lastCheckDateOption: Option[DateTime] =
     lastCheckMsOption map (lastCheckMs => new DateTime(lastCheckMs))
 }