+++ src/main/scala/org/fs/checker/notification/UpdateNotifierServiceImpl.scala
@@ -20,18 +20,32 @@ import org.slf4s.Logging
  * @author FS
  */
 class UpdateNotifierServiceImpl extends UpdateNotifierService with Logging {
-  override def notify(entries: Seq[TorrentEntry]): Unit =
-    notifyInner(entries)
+  override def notify(updated: Seq[TorrentEntry], notAvailable: Seq[(TorrentEntry, String)]): Unit = {
+    if (updated.nonEmpty || notAvailable.nonEmpty) {
+      notifyInner(updated, notAvailable)
+    }
+  }
 
-  def notifyInner(entries: Seq[TorrentEntry]): Future[Unit] = {
-    entries.foreach {
+  def notifyInner(updated: Seq[TorrentEntry], notAvailable: Seq[(TorrentEntry, String)]): Future[Unit] = {
+    updated.foreach {
       case TorrentEntry(alias, url) => log.warn(s"'$alias' ($url) was updated!")
+    }
+    notAvailable.foreach {
+      case (TorrentEntry(alias, url), reason) => log.warn(s"'$alias' ($url) is invalid: $reason!")
     }
 
     Future {
-      val linkLabels = entries map (e => new LinkLabel(e.alias, new URI(e.url)))
-      val commaLabels = Seq.fill(linkLabels.size - 1)(new Label(", "))
-      val labels = linkLabels.head +: commaLabels.zip(linkLabels.tail).flatMap(p => Seq(p._1, p._2)) :+ new Label(" updated!")
+      val updatedLabels = if (updated.nonEmpty) {
+        val linkLabels = updated map (e => new LinkLabel(e.alias, new URI(e.url)))
+        val commaLabels = Seq.fill(linkLabels.size - 1)(new Label(", "))
+        linkLabels.head +: commaLabels.zip(linkLabels.tail).flatMap(p => Seq(p._1, p._2)) :+ new Label(" updated! ")
+      } else Seq.empty
+
+      val notAvailableLabels = if (notAvailable.nonEmpty) {
+        val linkLabels = notAvailable map (e => new LinkLabel(e._1.alias, new URI(e._1.url)))
+        val commaLabels = Seq.fill(linkLabels.size - 1)(new Label(", "))
+        linkLabels.head +: commaLabels.zip(linkLabels.tail).flatMap(p => Seq(p._1, p._2)) :+ new Label(" unavailable!")
+      } else Seq.empty
 
       val title = "Torrents updated"
       // Frame is dummy constructed for the dialog to be shown on the Windows taskbar
@@ -47,7 +61,7 @@ class UpdateNotifierServiceImpl extends UpdateNotifierService with Logging {
       }
       frame.visible = true
       val content = new BorderPanel {
-        val inner = new FlowPanel(labels: _*) {
+        val inner = new FlowPanel((updatedLabels ++ notAvailableLabels): _*) {
           hGap = 1
           vGap = 1
         }
@@ -70,6 +84,8 @@ object UpdateNotifierServiceImpl extends App {
   val f = (new UpdateNotifierServiceImpl).notifyInner(Seq(
     TorrentEntry("проверка", "http://www.example.com"),
     TorrentEntry("google", "http://www.google.com")
+  ), Seq(
+    (TorrentEntry("qwe", "http://www.example.com"), "!11")
   ))
   Await.ready(f, 100.second)
 }