+++ src/main/scala/org/fs/checker/provider/impl/Alltor.scala
@@ -6,6 +6,7 @@ import java.nio.charset.StandardCharsets
 import scala.util.Try
 
 import org.apache.http.client.HttpClient
+import org.fs.checker.dao.TorrentParseResult
 import org.fs.checker.dumping.PageContentDumpService
 import org.fs.checker.provider.ConfiguredProvider
 import org.fs.checker.provider.GenProvider
@@ -33,7 +34,7 @@ class Alltor(httpClient: HttpClient, override val dumpService: PageContentDumpSe
     ResponseBodyDecoder.bodyToString(resp)
   }
 
-  override def parseDateLastUpdated(content: String): DateTime = {
+  override def parseDateLastUpdated(content: String): TorrentParseResult = {
     val body = parseElement(content) \ "body"
     val downloadTable = body \\ "table" filterByClass "dl_list"
     val infoNode = (downloadTable \ "tr")(1) \ "td"
@@ -41,7 +42,7 @@ class Alltor(httpClient: HttpClient, override val dumpService: PageContentDumpSe
     val lastUpdatedString = lastUpdatedNode.trimmedText
     val duration = DurationParser.parse(lastUpdatedString)
     val result = DateTime.now - duration
-    result
+    TorrentParseResult.Success(result)
   }
 }