+++ src/test/scala/org/fs/checker/provider/impl/NonameClubSpec.scala
@@ -5,6 +5,7 @@ import java.io.File
 import scala.io.Source
 
 import org.fs.checker.TestHelper
+import org.fs.checker.dao.TorrentParseResult
 import org.joda.time.Days
 import org.junit.runner.RunWith
 import org.scalatest.FlatSpec
@@ -24,7 +25,8 @@ class NonameClubSpec
   it should "parse 09 Jun 2019" in {
     val content = Source.fromFile(new File(pagesFolder, "elementary_2019-06-09.html"), "windows-1251").mkString
     val parsed = instance.parseDateLastUpdated(content)
-    assert(parsed === DateTime.parse("2019-06-09T14:11:53"))
+    assert(parsed.isInstanceOf[TorrentParseResult.Success])
+    assert(parsed.asInstanceOf[TorrentParseResult.Success].dt === DateTime.parse("2019-06-09T14:11:53"))
   }
 
   val pagesFolder: java.io.File = new File(resourcesFolder, instance.providerKey)