+++ src/test/scala/org/fs/checker/provider/impl/RuTrackerSpec.scala
@@ -5,6 +5,7 @@ import java.io.File
 import scala.io.Source
 
 import org.fs.checker.TestHelper
+import org.fs.checker.dao.TorrentParseResult
 import org.joda.time.Days
 import org.junit.runner.RunWith
 import org.scalatest.FlatSpec
@@ -24,7 +25,15 @@ class RuTrackerSpec
   it should "parse 10 Jun 2019" in {
     val content = Source.fromFile(new File(pagesFolder, "elementary_2019-06-10.htm"), "windows-1251").mkString
     val parsed = instance.parseDateLastUpdated(content)
-    assert(parsed === DateTime.parse("2019-06-10T20:39:00"))
+    assert(parsed.isInstanceOf[TorrentParseResult.Success])
+    assert(parsed.asInstanceOf[TorrentParseResult.Success].dt === DateTime.parse("2019-06-10T20:39:00"))
+  }
+
+  it should "parse absorbed" in {
+    val content = Source.fromFile(new File(pagesFolder, "the-boys_2020-10-19.htm"), "utf8").mkString
+    val parsed = instance.parseDateLastUpdated(content)
+    assert(parsed.isInstanceOf[TorrentParseResult.NotAvailable])
+    assert(parsed.asInstanceOf[TorrentParseResult.NotAvailable].reason === "absorbed")
   }
 
   val pagesFolder: java.io.File = new File(resourcesFolder, instance.providerKey)