+++ src/test/scala/org/fs/checker/provider/impl/RutorSpec.scala
@@ -5,6 +5,7 @@ import java.io.File
 import scala.io.Source
 
 import org.fs.checker.TestHelper
+import org.fs.checker.dao.TorrentParseResult
 import org.joda.time.Days
 import org.junit.runner.RunWith
 import org.scalatest.FlatSpec
@@ -24,7 +25,8 @@ class RutorSpec
   it should "parse 2019-02-28 state" in {
     val content = Source.fromFile(new File(pagesFolder, "young-sheldon_2019-02-28.html"), "UTF-8").mkString
     val parsed = instance.parseDateLastUpdated(content)
-    assert(parsed === DateTime.parse("2019-02-28T23:19:22"))
+    assert(parsed.isInstanceOf[TorrentParseResult.Success])
+    assert(parsed.asInstanceOf[TorrentParseResult.Success].dt === DateTime.parse("2019-02-28T23:19:22"))
   }
 
   val pagesFolder: java.io.File = new File(resourcesFolder, instance.providerKey)